Radar Tools: RCWL-0516 and Raspberry PI
Three applications running on a Raspberry PI interfacing
with the RCWL-0516 microwave Doppler proximity sensor.

Materials needed:
1 Raspberry PI
1 RCWL-0516
1 Raspberry PI Camera
1 4P Pin Header (optional)
3 Female to Female Jumper Wires

The files:
The radarImageRun.py will take still images of what the RCWL-0516
detected.
The radarMotionRun.py will take video of what the RCWL-0516 detected.
The radarTest.py will print a 1 to the console when motion is detected
by the RCWL-0516.

Modified the README.md file to include additional information
about this project.

	new file:   radarTools/radarImageRun.py
	new file:   radarTools/radarMotionRun.py
	new file:   radarTools/radarTest.py